# Conflicts:
#	Core/Compatibility.cpp
#	Core/Compatibility.h
#	Core/Config.cpp
#	Core/Config.h
#	Core/CoreTiming.cpp
#	Core/HLE/sceDisplay.cpp
#	GPU/Common/DrawEngineCommon.h
#	GPU/Common/SplineCommon.cpp
#	GPU/D3D11/DrawEngineD3D11.cpp
#	GPU/D3D11/ShaderManagerD3D11.cpp
#	GPU/Directx9/ShaderManagerDX9.cpp
#	GPU/Directx9/VertexShaderGeneratorDX9.cpp
#	GPU/GLES/DrawEngineGLES.cpp
#	GPU/GPUCommon.cpp
#	GPU/Vulkan/DrawEngineVulkan.cpp
#	UI/CwCheatScreen.cpp
#	UI/EmuScreen.cpp
#	UI/GameSettingsScreen.cpp
#	UI/MainScreen.cpp
#	Windows/GEDebugger/VertexPreview.cpp
#	Windows/MainWindowMenu.cpp
#	Windows/ppsspp.rc
#	Windows/resource.h
#	assets/compat.ini
#	ext/native/thin3d/VulkanRenderManager.cpp
#	libretro/libretro.cpp

+++ Windows/GEDebugger/VertexPreview.cpp
@@ -204,7 +204,7 @@ static void ExpandBezier(int &count, int op, const std::vector<SimpleVertex> &si
 	cpoints.col = (Vec4f *)AllocateAlignedMemory(sizeof(Vec4f) * num_points, 16);
 	cpoints.Convert(points.data(), num_points);
 
-	surface.Init(generatedVerts.size());
+	surface.Init((int)generatedVerts.size());
 	SoftwareTessellation(output, surface, gstate.vertType, cpoints);
 	count = output.count;
 
@@ -256,7 +256,7 @@ static void ExpandSpline(int &count, int op, const std::vector<SimpleVertex> &si
 	cpoints.col = (Vec4f *)AllocateAlignedMemory(sizeof(Vec4f) * num_points, 16);
 	cpoints.Convert(points.data(), num_points);
 
-	surface.Init(generatedVerts.size());
+	surface.Init((int)generatedVerts.size());
 	SoftwareTessellation(output, surface, gstate.vertType, cpoints);
 	count = output.count;
